Evaluate AST and generate the resulting Quilt in a matrix form.
While generating matrix from AST we declared string variables for matrix values and because of this when we update values in the matrix then Tokens are aﬀected. Matrix rotation was the another problem, it rotate the matrix every time ‘TURN’ operation occur. Third problem was the global declaration of string matrix which aﬀect the ‘SEW’ and ‘TURN’ operations.

Solution: To solve the string variable problem we declared enum datatype (var) and created a matrix class of var return type which is the base class of expnode. We declared the object of matrix class in each subclass of expnode. We also created a virtual evaluate function that returns a matrix. Evaluate function is used to evaluate AST and then generate the corresponding matrix. The ‘SEW’ and ‘TURN’ operations are identiﬁed by his parameter (e.g., TURN(A),SEW(A,B)). To print the matrix we declared a mprint () function inside the expnode. mprint() function is to evaluate enum datatype by mapping of integer value with every entity so if ‘TURN’ or ‘SEW’ operation is repeated then it increment the integer value and return the corresponding enum value. The matrix rotation problem is solved by the declaration of separate class for every operation of ‘Quilt’ and each have some same function (mprint(), evaluate(), print()).