Compile AST and generate C++ code to compute the resulting Quilt in a matrix form.
While compile AST and generate C++ code to compute the resulting Quilt in a matrix form we declared all the output ﬁles globally, so whenever we execute one command all the ﬁles generated automatically. Another problem was that when we call expnode in the yacc ﬁle with the the help of functions for all the commands then it generate a error message. The third problem was that when we use command line argument in yacc ﬁle then it can not count the number of commands argc(). 

Solution: We generated all ﬁles locally so when we call any function it will only generate its related ﬁle. We created a pointer for expnode in yacc ﬁle, by using that pointer we access all the functions which solved our second problem. In the main function to solve the problem of counting number of commands argc(), we use for loop from 0 to argc() and if- else statement inside the for loop to call the functions according to the command line.